feat: role distribution on Validator detail
