+++ apps/tangle-dapp/app/nomination/[validatorAddress]/ValidatorBasicInfoCard.tsx
@@ -9,13 +9,19 @@ import {
   Typography,
 } from '@webb-tools/webb-ui-components';
 import { shortenString } from '@webb-tools/webb-ui-components/utils/shortenString';
-import { FC } from 'react';
+import { FC, useEffect, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
 
 import { SocialChip, TangleCard } from '../../../components';
 import useNetworkStore from '../../../context/useNetworkStore';
-import { formatTokenBalance } from '../../../utils/polkadot';
-import useValidatorBasicInfo from './hooks/useValidatorBasicInfo';
+import useRestakingProfile from '../../../data/restaking/useRestakingProfile';
+import useCurrentEra from '../../../data/staking/useCurrentEra';
+import {
+  extractDataFromIdentityInfo,
+  formatTokenBalance,
+  IdentityDataType,
+} from '../../../utils/polkadot';
+import { getPolkadotApiPromise } from '../../../utils/polkadot/api';
 import ValueSkeleton from './ValueSkeleton';
 
 interface ValidatorBasicInfoCardProps {
@@ -27,7 +33,7 @@ const ValidatorBasicInfoCard: FC<ValidatorBasicInfoCardProps> = ({
   validatorAddress,
   className,
 }: ValidatorBasicInfoCardProps) => {
-  const { network, nativeTokenSymbol } = useNetworkStore();
+  const { network, nativeTokenSymbol, rpcEndpoint } = useNetworkStore();
 
   const {
     name,
@@ -39,7 +45,7 @@ const ValidatorBasicInfoCard: FC<ValidatorBasicInfoCardProps> = ({
     email,
     web,
     isLoading,
-  } = useValidatorBasicInfo(validatorAddress);
+  } = useValidatorBasicInfo(rpcEndpoint, validatorAddress);
 
   return (
     <TangleCard className={twMerge('min-h-[300px]', className)}>
@@ -152,3 +158,88 @@ const ValidatorBasicInfoCard: FC<ValidatorBasicInfoCardProps> = ({
 };
 
 export default ValidatorBasicInfoCard;
+
+function useValidatorBasicInfo(rpcEndpoint: string, validatorAddress: string) {
+  const { data: currentEra } = useCurrentEra();
+  const { profileTypeOpt, totalRestaked } =
+    useRestakingProfile(validatorAddress);
+
+  const [name, setName] = useState<string | null>(null);
+  const [email, setEmail] = useState<string | null>(null);
+  const [web, setWeb] = useState<string | null>(null);
+  const [twitter, setTwitter] = useState<string | null>(null);
+  const [nominations, setNominations] = useState<number | null>(null);
+  const [isActive, setIsActive] = useState<boolean | null>(null);
+  const [isLoading, setIsLoading] = useState(true);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      const api = await getPolkadotApiPromise(rpcEndpoint);
+      const fetchNameAndSocials = async () => {
+        const identityData = await api.query.identity.identityOf(
+          validatorAddress
+        );
+
+        if (identityData.isSome) {
+          const identity = identityData.unwrap();
+          const info = identity[0]?.info;
+          if (info) {
+            setName(extractDataFromIdentityInfo(info, IdentityDataType.NAME));
+            setEmail(extractDataFromIdentityInfo(info, IdentityDataType.EMAIL));
+            setWeb(extractDataFromIdentityInfo(info, IdentityDataType.WEB));
+            const twitterName = extractDataFromIdentityInfo(
+              info,
+              IdentityDataType.TWITTER
+            );
+            setTwitter(
+              twitterName === null
+                ? null
+                : `https://twitter.com/${twitterName.substring(1)}`
+            );
+          }
+        }
+      };
+
+      const fetchNominations = async () => {
+        if (currentEra === null || !api.query.staking?.erasStakersOverview) {
+          setNominations(null);
+          setIsActive(null);
+          return;
+        }
+
+        const erasStakersOverviewData =
+          await api.query.staking.erasStakersOverview(
+            currentEra,
+            validatorAddress
+          );
+        if (erasStakersOverviewData.isSome) {
+          const nominatorCount =
+            erasStakersOverviewData.unwrap().nominatorCount;
+          setNominations(nominatorCount.toNumber());
+          setIsActive(true);
+          return;
+        }
+
+        setNominations(null);
+        setIsActive(false);
+      };
+
+      await Promise.all([fetchNameAndSocials(), fetchNominations()]);
+      setIsLoading(false);
+    };
+
+    fetchData();
+  }, [validatorAddress, rpcEndpoint, currentEra]);
+
+  return {
+    name,
+    isActive,
+    totalRestaked,
+    restakingMethod: profileTypeOpt,
+    nominations,
+    twitter,
+    email,
+    web,
+    isLoading,
+  };
+}
